Python 3 compatibility: don't use tuple unpacking in lambdas.
Python 3 got rid of tuple unpacking in function arguments, so we have to use traditional indexing instead.

(I'm considering changing ttftree.CompoundMeasure to accept a custom tuple class to use; then I could pass in a custom namedtuple with "index" and "key" attributes.)

We're now syntax compatible with Python 3, but a few more things (such as calls to dict.iteritems) need to be changed before we can do anything useful. I'll be using six for that.