A few things:
Change TWeakRef to create the underlying weak reference inside __init__. This fixes a bug where creating a TWeakRef to an unweakrefable value (like an int) would cause breakage inside _make_mature.

Get rid of support for TWeakRef(None). The code for making this work was becoming burdensome, and it'll be a lot easier to handle this sort of thing in a utility wrapper class.

Add is_alive, which just returns self.get() is not None. This will allow for API compatibility with a utility wrapper class that allows None values.